feat: implement bius logic
