+++ frontend/appflowy_flutter/lib/plugins/document/presentation/editor_plugins/plugins.dart
@@ -29,9 +29,9 @@ export 'link_preview/link_preview_cache.dart';
 export 'link_preview/link_preview_menu.dart';
 export 'math_equation/math_equation_block_component.dart';
 export 'math_equation/mobile_math_equation_toolbar_item.dart';
+export 'mobile_toolbar_v3/aa_toolbar_item.dart';
 export 'mobile_toolbar_v3/appflowy_mobile_toolbar.dart';
 export 'mobile_toolbar_v3/appflowy_mobile_toolbar_item.dart';
-export 'mobile_toolbar_v3/text_decoration_item.dart';
 export 'openai/widgets/auto_completion_node_widget.dart';
 export 'openai/widgets/smart_edit_node_widget.dart';
 export 'openai/widgets/smart_edit_toolbar_item.dart';
